requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
pymysql>=1.1.0
python-multipart>=0.0.9
jinja2>=3.1.2